+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @openfn/cli
 
+## 1.20.0
+
+### Minor Changes
+
+- 69ec22a: Refactor of openfn project command. There are very few user-facing changes, and they should be compatible
+
+  - A new `project` namespace has been set up, allowing `openfn project version|list|merge|checkout`
+  - `openfn projects` will continue to list projects in the workspace (but is just an alias of list)
+  - The prior `openfn merge|checkout` command still exist, it just aliases to `openfn projct merge|checkout`
+
+  One change to watch out for is that `--project-path` has been changed to `--workspace`, which can also be set through `-w` and `OPENFN_WORKSPACE`.
+
+- 162e0ea: Add a fetch command, which will download a project from an app but not check it out. This will throw if the local project version has diverged from the remote version.
+
+  Rebased `pull --beta` to simply be fetch & checkout
+
 ## 1.19.0
 
 ### Minor Changes